+++ src/main/java/zerobase/matching/comment/dto/ReadComment.java
@@ -11,14 +11,14 @@ public class ReadComment {
   @Getter // 이거 없으면 403 에러 뜸 ( 클라이언트가 요청한 Type 으로 응답을 내려줄 수 없다 )
   public static class Response {
 
-    private long commentId;
-    private long projectId;
-    private long userId;
+    private int commentId;
+    private int projectId;
+    private int userId;
     private String content;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
     private LocalDateTime deleteTime;
-    private long parentId;
+    private int parentId;
     private int level;
 
     public static ReadComment.Response fromEntity(CommentDto commentDto) {
